requirements.txt
anyio==4.9.0
certifi==2025.1.31
distlib==0.3.8
exceptiongroup==1.2.2
filelock==3.13.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
platformdirs==4.2.0
python-telegram-bot==22.0
sniffio==1.3.1
typing_extensions==4.13.0
virtualenv==20.25.1
